requirements.txt
# Core UI + extraction
streamlit>=1.24.0
PyMuPDF>=1.22.0

# OCR / vision
easyocr>=1.6.0
Pillow>=10.0.0

# ML / embeddings
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.34.0
accelerate>=0.21.0
faiss-cpu>=1.7.3

# Language utilities
numpy>=1.26.0
langdetect>=1.0.9

# Hugging Face model download + safetensors support
huggingface-hub>=0.16.0
safetensors>=0.4.0

# BGE-M3 embeddings
sentence-transformers
numpy